+++ Editor/Generatable/GeneratableComment.cs
@@ -1,0 +1,11 @@
+namespace NPTP.UnitySourceGen.Editor.Generatable
+{
+    public class GeneratableComment : GeneratableBase
+    {
+        public GeneratableComment(string comment) : base(comment, default, default)
+        {
+        }
+
+        public override string GenerateStringRepresentation() => Name; // Name == comment
+    }
+}