+++ Editor/Generatable/GeneratableProperty.cs
@@ -1,10 +1,9 @@
 using NPTP.UnitySourceGen.Editor.Enums;
-using NPTP.UnitySourceGen.Editor.Options;
 
 namespace NPTP.UnitySourceGen.Editor.Generatable
 {
     public abstract class GeneratableProperty : GeneratableBase
     {
-        protected GeneratableProperty(NameSyntax nameSyntax, AccessModifier accessModifier, bool isStatic) : base(nameSyntax, accessModifier, isStatic) { }
+        protected GeneratableProperty(string name, AccessModifier accessModifier, bool isStatic) : base(name, accessModifier, isStatic) { }
     }
 }