+++ Editor/Generatable/GeneratableTypeDefinition.cs
@@ -22,7 +22,7 @@ namespace NPTP.UnitySourceGen.Editor.Generatable
         private List<GeneratableProperty> Properties { get; } = new();
         private List<GeneratableMethod> Methods { get; } = new();
 
-        internal GeneratableTypeDefinition(string nameSyntax, AccessModifier accessModifier, bool isStatic) : base(nameSyntax, accessModifier, isStatic) { }
+        internal GeneratableTypeDefinition(string name, AccessModifier accessModifier, bool isStatic) : base(name, accessModifier, isStatic) { }
 
         public override string GenerateStringRepresentation()
         {